fold the location handling from Site into a generic base class
