+++ lib/models/activity.js
@@ -19,6 +19,8 @@ class Activity extends fb.FirebaseRecord {
      * NOTE: Adding a location is a separate write operation to saving or updating the
      *       main Activity record.
      *
+     * @todo Wrap the double write in a transaction.
+     *
      * @param {array} location - An array containing a `[latitude, longitude]` pair.
      * @param {Site} site - The site this location is associated with.
      * @return {Promise} The operation to write the location data.
@@ -39,20 +41,6 @@ class Activity extends fb.FirebaseRecord {
 class Activities extends fb.FirebaseCollection {
 
     /**
-     * Identical to `FirebaseCollection#newRecord` but returns {@link Activity} instances.
-     */
-    newRecord() {
-        if (arguments.length == 0) {
-            return new Activity(this.root.push(), {})
-        }
-        else if (arguments.length == 1) {
-            return new Activity(this.root.push(), arguments[0]);
-        } else {
-            return new Activity(this.root.child(arguments[0]), arguments[1]);
-        }
-    }
-
-    /**
      * Deletes all Activity records as well as the locations collection.
      */
     drop() {
@@ -60,4 +48,6 @@ class Activities extends fb.FirebaseCollection {
     }
 }
 
-module.exports = new Activities(host.child('activities'));
+const activities = new Activities(host.child('activities'));
+activities.recordClass = Activity;
+module.exports = activities;
