+++ lib/models/firebase-classes.js
@@ -13,6 +13,7 @@
  * aces.write();
  */
 'use strict';
+const GeoFire = require('geofire');
 
 /** A private symbol for classes to store their underlying reference under. */
 const refKey = Symbol('ref');
@@ -82,6 +83,58 @@ class FirebaseRecord {
 }
 
 /**
+ * A basic extension of {@link FirebaseRecord} that looks for the presence of a 'location'
+ * property and saves that data via GeoFire instead.
+ *
+ * @todo Wrap the double write in a transaction.
+ */
+class LocationAwareRecord extends FirebaseRecord {
+
+    /**
+     * @return {GeoFire} The GeoFire wrapper around the base reference.
+     */
+    get geofire() {
+        if (!this[geoRef]) {
+            this[geoRef] = new GeoFire(this.ref);
+        }
+        return this[geoRef];
+    }
+
+    /**
+     * If location data is present first writes all other data normally then
+     * uses the `geofire` wrapper to set the location data.
+     */
+    write() {
+        let location = this.location;
+        if (location) {
+            delete this.location;
+            return super.write().then(ok => {
+                    return this.geofire.set('location', location);
+                });
+        }
+        return super.write();
+    }
+
+    /**
+     * If location data is present first updates all other data normally then
+     * uses the `geofire` wrapper to set the location data.
+     */
+    update() {
+        let location = this.location;
+        if (location) {
+            delete this.location;
+            return super.update().then(ok => {
+                    return this.geofire.set('location', location);
+                });
+        }
+        return super.update();
+    }
+}
+
+/** A private symbol for storing the GeoFire reference wrapper under */
+const geoRef = Symbol('geofire');
+
+/**
  * Acts as a collection of data in Firebase, can be used to create new {@link FirebaseRecord}
  * instances for saving.
  *
@@ -93,16 +146,16 @@ class FirebaseCollection {
      * Creates the collection at a root Firebase reference.
      *
      * @param {Firebase} root - The underlying reference at which all new records should be created.
+     * @param {function} recordClass - A `class` or function to be used as the base of all created records,
+     *                                 defaults to {@link FirebaseRecord} if not set.
      */
-    constructor(root) {
-        this[refKey] = root;
-    }
-
-    /**
-     * @return {Firebase} The underlying Firebase reference which is the root of the collection.
-     */
-    get root() {
-        return this[refKey];
+    constructor(root, recordClass) {
+        this.root = root;
+        if (recordClass) {
+            this.recordClass = recordClass;
+        } else {
+            this.recordClass = FirebaseRecord;
+        }
     }
 
     /**
@@ -125,12 +178,12 @@ class FirebaseCollection {
      */
     newRecord() {
         if (arguments.length == 0) {
-            return new FirebaseRecord(this.root.push(), {})
+            return new this.recordClass(this.root.push(), {})
         }
         else if (arguments.length == 1) {
-            return new FirebaseRecord(this.root.push(), arguments[0]);
+            return new this.recordClass(this.root.push(), arguments[0]);
         } else {
-            return new FirebaseRecord(this.root.child(arguments[0]), arguments[1]);
+            return new this.recordClass(this.root.child(arguments[0]), arguments[1]);
         }
     }
 
@@ -147,5 +200,6 @@ class FirebaseCollection {
 
 module.exports = {
     FirebaseRecord: FirebaseRecord,
+    LocationAwareRecord: LocationAwareRecord,
     FirebaseCollection: FirebaseCollection
 };
